Update PopMenuViewController.swift
Remove @objc modifiers. They cause mixed objc/swift projects to fail to compile because any class that uses the delegate won't be able to find the delegate definition in the bridge file. 

This is because the protocol passes PopMenuViewController but that class is not @objc compatible.  

Remove the @objc, remove the optional because it requires objc. It's irrelevant anyway because there is only one function, therefore anyone passing in the delegate should implement that function